Split files, Created Folder
