+++ src/components/MarketWidget/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Index file for MarketWidget components
+ * 
+ * This file makes importing these components easier by providing a single entry point.
+ * Instead of:
+ *   import { MarketsWidget } from '@/components/MarketWidget/MarketsWidget';
+ * 
+ * You can use:
+ *   import { MarketsWidget } from '@/components/MarketWidget';
+ */
+export { default as MarketsWidget } from './MarketsWidget';
+export { MarketsWidgetWrapper } from './MarketsWidgetWrapper';
+export { MarketsWidgetHeader } from './MarketsWidgetHeader';
+export { MarketsWidgetMenu } from './MarketsWidgetMenu';
+export { MarketsWidgetFilter } from './MarketsWidgetFilter';
